Adds max cluster points option
Implements a feature to limit the number of points displayed in a cluster.

This enhances the visual representation of clustered data by preventing overcrowding and improving readability. It allows to specify a maximum number of points to display in a cluster, showing a "+"-sign if the actual number of points exceeds this limit.

Adds a parameter to the cluster definition to allow this configuration.